+++ R/biproportional-wto.R
@@ -13,8 +13,8 @@ create_wto_round_function = function(votes_matrix, district_seats, seats_parties
     assert_no_duplicates(names(district_seats))
     assert_no_duplicates(names(seats_parties))
 
-    stopifnot(identical(colnames(votes_matrix), names(district_seats)))
-    stopifnot(identical(rownames(votes_matrix), names(seats_parties)))
+    assert(identical(colnames(votes_matrix), names(district_seats)))
+    assert(identical(rownames(votes_matrix), names(seats_parties)))
 
     DISTRICT_WINNERS = most_votes_in_district_matrix(votes_matrix)